+++ src/export/resume_helpers.py
@@ -1,60 +1,21 @@
+# src/export/resume_helpers.py
 from __future__ import annotations
 
 from datetime import datetime
-from typing import Any, Optional, List, Dict
+from typing import Any, Dict, List
 
 from docx import Document
-import re
+
+from .shared_helpers import (
+    parse_date,
+    format_date_range,
+    strip_percent_tokens,
+    clean_languages_above_threshold,
+)
 
 # -------------------------
-# Date helpers
+# Resume-only helpers (DOCX)
 # -------------------------
-
-def parse_date(value: Any) -> Optional[datetime]:
-    """
-    Best-effort parse for common DB / snapshot date strings.
-    Supports:
-      - YYYY-MM-DD
-      - YYYY-MM-DD HH:MM:SS
-      - YYYY-MM-DDTHH:MM:SS
-    """
-    if not isinstance(value, str):
-        return None
-    s = value.strip()
-    if not s:
-        return None
-
-    candidates = [s, s[:19]]
-    fmts = ("%Y-%m-%d", "%Y-%m-%d %H:%M:%S", "%Y-%m-%dT%H:%M:%S")
-    for c in candidates:
-        for fmt in fmts:
-            try:
-                return datetime.strptime(c, fmt)
-            except ValueError:
-                continue
-    return None
-
-
-def format_date_range(start: Any, end: Any) -> str:
-    """
-    Output examples:
-      - 'Nov 2024 – Dec 2024'
-      - 'Sep 2024 – Present'
-      - '' if no dates
-    """
-    ds = parse_date(start)
-    de = parse_date(end)
-
-    def fmt(d: datetime) -> str:
-        return d.strftime("%b %Y")
-
-    if ds and de:
-        return f"{fmt(ds)} – {fmt(de)}"
-    if ds and not de:
-        return f"{fmt(ds)} – Present"
-    if not ds and de:
-        return fmt(de)
-    return ""
 
 def add_role_date_line(doc: Document, role: str, date_line: str) -> None:
     """
@@ -78,6 +39,7 @@ def add_role_date_line(doc: Document, role: str, date_line: str) -> None:
         if p.runs:
             p.runs[0].italic = True
 
+
 def _project_sort_key(p: dict) -> datetime:
     """
     Sort by:
@@ -96,12 +58,9 @@ def _project_sort_key(p: dict) -> datetime:
     # projects with no dates go last
     return datetime.min
 
-# -------------------------
-# DOCX helpers
-# -------------------------
 
 def add_section_heading(doc: Document, title: str) -> None:
-    doc.add_heading(title.upper(), level=1)
+    doc.add_heading((title or "").upper(), level=1)
 
 
 def add_placeholder(doc: Document, text: str) -> None:
@@ -115,49 +74,16 @@ def add_bullet(doc: Document, text: str) -> None:
     p.paragraph_format.space_after = 0
 
 
-# Clean language in skills section
-
-_LANG_PCT_RE = re.compile(r"^\s*(?P<name>.+?)\s+(?P<pct>\d+)\s*%\s*$")
-
-def clean_languages_above_threshold(
-    values: Any,
-    *,
-    min_pct: int = 10,
-) -> List[str]:
-    """
-    Input example:
-      ["Python 88%", "JavaScript 54%", "CSS 10%", "JSON 5%", "JSON 6%"]
-    Output (min_pct=10, strict >):
-      ["Python", "JavaScript"]  # (and other >10 items)
-    Rules:
-      - strips percentages entirely
-      - keeps only pct > min_pct
-      - dedupes by taking max pct per language
-      - ignores malformed entries
-    """
-    if not isinstance(values, list):
-        return []
-
-    best: Dict[str, int] = {}
-
-    for raw in values:
-        s = str(raw).strip()
-        if not s:
-            continue
-
-        m = _LANG_PCT_RE.match(s)
-        if not m:
-            continue
-
-        name = m.group("name").strip()
-        pct = int(m.group("pct"))
-
-        if pct < min_pct:  # strict ">"
-            continue
-
-        prev = best.get(name)
-        if prev is None or pct > prev:
-            best[name] = pct
-
-    # sort by pct desc, then name asc for stable output
-    return [name for name, _pct in sorted(best.items(), key=lambda kv: (-kv[1], kv[0].lower()))]
+__all__ = [
+    # shared helpers
+    "parse_date",
+    "format_date_range",
+    "strip_percent_tokens",
+    "clean_languages_above_threshold",
+    # resume helpers
+    "add_role_date_line",
+    "_project_sort_key",
+    "add_section_heading",
+    "add_placeholder",
+    "add_bullet",
+]
